requirements.txt
blinker==1.9.0
click==8.1.8
ffmpeg-python==0.2.0
Flask==3.1.0
future==1.0.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
pytest==8.3.4
Werkzeug==3.1.3
